CVS: ---------------------------------------------------------------------- CVS: Please format log message lines to be no longer than the above line. MSVC++ 2005 take 2. More warnings quelched.
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1275 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/engine/map_engine.cpp
@@ -3519,7 +3519,7 @@ CMapEngine::OpenMap(const char* filename)
     }
 
     // reset all the layer angles
-    for (i = 0; i < m_Map.GetMap().GetNumLayers(); i++)
+    for (i = 0; i < (unsigned int)m_Map.GetMap().GetNumLayers(); i++)
     {
         m_Map.SetLayerAngle(i, 0.0);
     }
